int update

+++ assign4.pde
@@ -121,8 +121,8 @@ void alienMaker(int num,int numInRow) {
 
   for (int i=0;i<num;i++){
     
-    int row=i/numInRow;
-    int col=i%numInRow;
+    int row=int(i/numInRow);
+    int col=int(i%numInRow);
     int x=50+(40*col);
     int y=50+(50*row);
     aList[i]= new Alien(x, y);
